# code/backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.ChangePasswordRequest:
    properties:
      confirm_password:
        type: string
      new_password:
        type: string
      old_password:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      profile_picture:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: info@dotsem.be
    name: Sem Van Broekhoven
    url: https://github.com/Thomas-More-Digital-Innovation/2526-DI-004-GoStrategy
  description: This is the API server for the GoStrategy game.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
  termsOfService: http://swagger.io/terms/
  title: GoStrategy API
  version: 0.1.1
paths:
...
  /users/{id}:
    get:
      description: Retrieve information about a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid or missing user ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user by ID
      tags:
      - users
...
  /users/me:
    get:
      description: Retrieve information about the currently logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current user
      tags:
      - users
...
  /users/me/password:
    post:
      consumes:
      - application/json
      description: Update the password for the currently logged-in user
      parameters:
      - description: Password update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or passwords don't match
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized or invalid current password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update password
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Change password
      tags:
      - users
...
swagger: "2.0"